requirements.txt
py-clob-client==0.34.5
websockets>=12.0
aiohttp>=3.9
python-dotenv>=1.0
pyyaml>=6.0
